updated controller to use new forDates service

+++ app/services/ExpensesService.scala
@@ -14,8 +14,11 @@ trait ExpensesService {
   def forDates(ExpensesQuery: ExpensesQuery): Future[Seq[CompanyCost]]
 }
 
-class RepositoryExpensesService(expensesRepo: ExpensesRepository, vat: Double)(implicit val ex: ExecutionContext) extends ExpensesService {
+class RepositoryExpensesService(expensesRepo: ExpensesRepository, vat: BigDecimal)
+                               (implicit val ex: ExecutionContext) extends ExpensesService {
   def save(expenses: Seq[Expense]): Future[Unit] = expensesRepo.save(expenses)
   def forDatesOld(ExpensesQuery: ExpensesQuery): Future[Seq[Expense]] = expensesRepo.forDates(ExpensesQuery)
-  def forDates(ExpensesQuery: ExpensesQuery): Future[Seq[CompanyCost]] = expensesRepo.forDates(ExpensesQuery).map { _.map(Converters.toCompanyCostFromExpenseWithVatRate(vat)) }
+  def forDates(ExpensesQuery: ExpensesQuery): Future[Seq[CompanyCost]] = expensesRepo.forDates(ExpensesQuery).map {
+    _.map(Converters.toCompanyCostFromExpenseWithVatRate(vat))
+  }
 }
