+++ app/services/package.scala
@@ -4,7 +4,7 @@ package object services {
   trait ServicesModule extends repos.ReposModule {
 	  import com.softwaremill.macwire._
 
-    val vatRate: Double
+    val vatRate: BigDecimal
     val ex: ExecutionContext
     lazy val expensesService = wire[RepositoryExpensesService]
   }
